Client: separated button definitions into a separate file
